extract InserWriter
